Combined map functionality and scoring with toolbar
