+++ lib/main.dart
@@ -1,7 +1,4 @@
-import 'dart:async';
 import 'package:flutter/material.dart';
-import 'package:google_maps_flutter/google_maps_flutter.dart';
-import 'dart:math';
 import 'package:umass_geoguessr_app/gamePage.dart';
 
 void main() {
@@ -60,137 +57,8 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
-  final Completer<GoogleMapController> _controller =
-      Completer<GoogleMapController>();
-
-  static const CameraPosition umasscampus = CameraPosition(
-    target: LatLng(42.390663, -72.528478),
-    zoom: 15.8,
-  );
-
-  LatLng guessPos = const LatLng(0, 0);
-
-  LatLng answerPos = const LatLng(42.389777, -72.523340);
-
-  Polyline line = const Polyline(
-      polylineId: PolylineId('guessanswerline'),
-      points: <LatLng>[LatLng(0, 0), LatLng(0, 0)]);
-
-  double dist = 0;
-
-  double degToRad(deg) {
-    return (deg * pi) / 180;
-  }
-
-  void makeGuess() {
-    setState(() {
-      // This call to setState tells the Flutter framework that something has
-      // changed in this State, which causes it to rerun the build method below
-      // so that the display can reflect the updated values. If we changed
-      // _counter without calling setState(), then the build method would not be
-      // called again, and so nothing would appear to happen.
-      line = Polyline(
-          width: 3,
-          polylineId: const PolylineId('guessanswerline'),
-          points: <LatLng>[guessPos, answerPos]);
-
-      // dist = acos(sin(lat1)*sin(lat2)+cos(lat1)*cos(lat2)*cos(lon2-lon1))*6371  (in kilometers)
-      dist = acos((sin(degToRad(guessPos.latitude)) *
-                  sin(degToRad(answerPos.latitude))) +
-              (cos(degToRad(guessPos.latitude)) *
-                  cos(degToRad(answerPos.latitude)) *
-                  cos(degToRad(answerPos.longitude - guessPos.longitude)))) *
-          6371;
-    });
-  }
-
-  void markGuess(coords) {
-    setState(() {
-      guessPos = coords;
-    });
-  }
-
   @override
   Widget build(BuildContext context) {
-    // This method is rerun every time setState is called, for instance as done
-    // by the _incrementCounter method above.
-    //
-    // The Flutter framework has been optimized to make rerunning build methods
-    // fast, so that you can just rebuild anything that needs updating rather
-    // than having to individually change instances of widgets.
-    return Scaffold(
-      appBar: AppBar(
-        // TRY THIS: Try changing the color here to a specific color (to
-        // Colors.amber, perhaps?) and trigger a hot reload to see the AppBar
-        // change color while the other colors stay the same.
-        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
-        // Here we take the value from the MyHomePage object that was created by
-        // the App.build method, and use it to set our appbar title.
-        title: Text(widget.title),
-      ),
-      body: Center(
-        // Center is a layout widget. It takes a single child and positions it
-        // in the middle of the parent.
-        child: Row(
-          // Column is also a layout widget. It takes a list of children and
-          // arranges them vertically. By default, it sizes itself to fit its
-          // children horizontally, and tries to be as tall as its parent.
-          //
-          // Column has various properties to control how it sizes itself and
-          // how it positions its children. Here we use mainAxisAlignment to
-          // center the children vertically; the main axis here is the vertical
-          // axis because Columns are vertical (the cross axis would be
-          // horizontal).
-          //
-          // TRY THIS: Invoke "debug painting" (choose the "Toggle Debug Paint"
-          // action in the IDE, or press "p" in the console), to see the
-          // wireframe for each widget.
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            Text(dist.toString()),
-            SizedBox(
-              width: 750,
-              height: 750,
-              child: GoogleMap(
-                onTap: (coords) => markGuess(coords),
-                markers: {
-                  Marker(
-                    markerId: const MarkerId('guess'),
-                    position: guessPos,
-                  ),
-                  Marker(
-                    markerId: const MarkerId('answer'),
-                    position: answerPos,
-                  )
-                },
-                polylines: {line},
-                cloudMapId: '74ae03550584037b',
-                cameraTargetBounds: CameraTargetBounds(LatLngBounds(
-                    southwest: const LatLng(42.383764, -72.536265),
-                    northeast: const LatLng(42.396610, -72.516493))),
-                minMaxZoomPreference: const MinMaxZoomPreference(15.5, 19),
-                initialCameraPosition: umasscampus,
-                onMapCreated: (GoogleMapController controller) {
-                  _controller.complete(controller);
-                },
-              ),
-            ),
-            Image.asset('images/Image 1.png', scale: 2.5),
-          ],
-        ),
-      ),
-      bottomNavigationBar: BottomAppBar(
-        shape: const CircularNotchedRectangle(),
-        child: Container(height: 50.0),
-      ),
-      floatingActionButton: FloatingActionButton.extended(
-        onPressed: makeGuess,
-        tooltip: 'Make your guess',
-        label: const Text('Guess'),
-        icon: const Icon(Icons.check_circle_outline),
-      ), // This trailing comma makes auto-formatting nicer for build methods.
-      floatingActionButtonLocation: FloatingActionButtonLocation.centerDocked,
-      home: const GamePage(title: 'UMass GeoGuessr'),
-    );
+    return const Scaffold(body: GamePage(title: 'UMass GeoGuessr'));
   }
 }
